refactor: split identifiers.go into hashing.go/transactions.go, where they are used
